split up game modules
