+++ src/app/allocations/Allocations.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { useSelector } from 'react-redux';
 
-import { ActivityType } from 'src/game';
+import { ActivityType } from 'src/game/types';
 import { gameStateSelector } from 'src/state';
 
 import { Allocation } from './AllocationItem';
