+++ src/game/constants.ts
@@ -1,0 +1,15 @@
+export const emptyResources = {
+  people: 0,
+  food: 0,
+  water: 0,
+  wood: 0,
+};
+
+export const emptyActivities = {
+  gather_lumber: 0,
+  defend: 0,
+  gather_food: 0,
+  hunt: 0,
+  collect_water: 0,
+  scout: 0,
+};
